backend/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
pkg_resources==0.0.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
